+++ CHANGELOG.rst
@@ -5,7 +5,7 @@ Changelog
 ------------------
 
 - Fix bug in tabs widget when using newer versions of jQuery 
-- Remove support for Django 1.7 and add support for Django 1.10
+- Add support for Django 1.11 and remove support for all versions prior.
 
 
 0.8.4 (2016-06-03)